requirements.txt
pandas
seaborn
scikit-learn
xgboost
mlflow
hydra-core
joblib
pytest
pytest-cov
bentoml
dvc
dvc-s3
black
fastapi
uvicorn
GitPython
requests
pyyaml
numpy
dagshub
pytest-html
requests 
pyyaml
pandera
